Move datasets to a file
