--- src/moscot/datasets/__init__.py
@@ -1,1 +1,0 @@
-from moscot.datasets._datasets import *  # noqa: F403
